4/study-hybrid-search-vertex/pyproject.toml
[project]
name = "study-hybrid-search-vertex"
version = "0.1.0"
description = "Hybrid search (Meilisearch BM25 + BigQuery VECTOR_SEARCH + RRF) with LightGBM LambdaRank rerank, migrating to Vertex AI"
requires-python = ">=3.12"
dependencies = [
    # --- shared (ml.common / ml.registry / ml.data) ---
    "pydantic>=2.8",
    "pydantic-settings[yaml]>=2.4",
    "google-cloud-bigquery>=3.25",
    "google-cloud-bigquery-storage>=2.27",
    "google-cloud-storage>=2.18",
    "google-cloud-logging>=3.11",
    "pandas>=2.2",
    "pyarrow>=16",
    "db-dtypes>=1.3",
    # --- app (FastAPI / Vertex AI / Pub/Sub) ---
    "cachetools>=5.3",
    "fastapi>=0.115",
    "google-auth>=2.36",
    "google-cloud-aiplatform>=1.71",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "numpy>=2.0",
    "google-cloud-pubsub>=2.23",
    "google-cloud-run>=0.10",
    # --- ml serving / training ---
    "lightgbm>=4.5",
    "sentence-transformers>=3.0",
    "wandb>=0.17",
    # --- pipelines (KFP / Vertex Pipelines) ---
    "google-cloud-pipeline-components>=2.17",
    "kfp>=2.9,<3",
]

[project.scripts]
rank-train = "ml.training.trainer:main"
embed = "pipeline.data_job.main:main"
meili-sync = "scripts.local.ops.sync_meili:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "ml", "pipeline"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "types-requests",
    # httpx powers fastapi.testclient (starlette); needed by tests/unit/app.
    "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
# UP017: keep `timezone.utc` so code imports cleanly on 3.10/3.11 local envs (target is 3.12).
# RUF012: SEVERITY_MAP on CloudLoggingJsonFormatter is intentionally class-level constant.
ignore = ["E501", "UP017", "RUF012"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra --strict-markers"
